Move page data fetching to separate function
